refactor: move microk8s cluster nginx code to a dedicated file
For better modularity and readability.
